@@ -1,24 +1,27 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import './index.scss';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import 'bootstrap/dist/js/bootstrap.bundle.min';
 import App from './components/app/App';
-import Projects from './components/projects/Projects';
+import ContactUs from './components/contactUs/ContactUs';
 import Bulletin from './components/bulletin/Bulletin';
-import { BrowserRouter, Route } from 'react-router-dom'
+import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import Topbar from './components/topbar/Topbar';
 import Footer from './components/footer/Footer';
 
+const container = document.getElementById('root');
+const root = createRoot(container);
 
-ReactDOM.render(
-    <BrowserRouter>
-    <Topbar/>
-      <Route exact path="/" component={App}/>
-      <Route exact path="/home" component={App}/>
-      <Route exact path="/contact_us" component={Projects}/>
-      <Route exact path="/bulletin" component={Bulletin}/>
-    <Footer/>
-    </BrowserRouter>,
-  document.getElementById('root')
+root.render(
+  <BrowserRouter>
+    <Topbar />
+    <Routes>
+      <Route path="/" element={<App />} />
+      <Route path="/home" element={<App />} />
+      <Route path="/contact_us" element={<ContactUs />} />
+      <Route path="/bulletin" element={<Bulletin />} />
+    </Routes>
+    <Footer />
+  </BrowserRouter>
 );
