@@ -11,20 +11,18 @@ import {
 import LinksListItem from "../linksListItem/LinksListItem";
 
 export default function Links() {
-// eslint-disable-next-line
+  // eslint-disable-next-line
   const [data, setData] = useState(linksData);
 
   return (
     <>
-    <Container className="links content-container pb-4" id="projects">
-      <Row className="mx-4">
-        {data.map((item) => (
-          <>
-            <LinksListItem item={item}/>
-          </>
-        ))}
-      </Row>
-    </Container>
+      <Container className="links content-container py-4" id="projects">
+        <Row className="links-row d-flex justify-content-center">
+          {data.map((item, index) => (
+            <LinksListItem key={item.id || index} item={item} />
+          ))}
+        </Row>
+      </Container>
     </>
   )
 }
